Replace theme toggle with react-toggle-dark-mode
