+++ index.html
@@ -76,7 +76,7 @@
                             <a class="nav-link" href="#contact"><i class="fas fa-envelope-open"></i><span class="nav-text">Contact</span></a>
                         </li>
                         <li class="nav-item ms-3 d-flex align-items-center">
-                            <button id="theme-toggle-btn" class="theme-toggle" aria-label="Toggle theme"><span class="icon"></span></button>
+                            <div id="dark-mode-toggle"></div>
                         </li>
                     </ul>
                 </div>
@@ -427,14 +427,7 @@
     <script src="js/components/dynamic-island.js"></script>
     <!-- Utils JS -->
     <script src="js/utils/storage.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react/17.0.2/umd/react.development.js" crossorigin></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/react-dom/17.0.2/umd/react-dom.development.js" crossorigin></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/babel-standalone/7.23.2/babel.min.js" crossorigin></script>
-    <script src="https://unpkg.com/@emotion/react@11.11.1/dist/emotion-react.umd.min.js" crossorigin></script>
-    <script src="https://unpkg.com/@emotion/styled@11.11.0/dist/emotion-styled.umd.min.js" crossorigin></script>
-    <script src="https://unpkg.com/@mui/material@5.15.8/umd/material-ui.development.js" crossorigin></script>
-    <script type="text/babel" src="js/react/components/VisitCounter.jsx"></script>
-    <script type="text/babel" src="js/react/main.jsx"></script>
+    <script type="module" src="js/react/main.mjs"></script>
     <script src="https://cdn.jsdelivr.net/npm/feather-icons/dist/feather.min.js" defer></script>
 </body>
